Separate runtime and compiler implementations
