+++ apywire/__init__.py
@@ -4,24 +4,25 @@
 
 """A package to wire up objects."""
 
+from .compiler import WiringCompiler
 from .exceptions import (
     CircularWiringError,
     LockUnavailableError,
     UnknownPlaceholderError,
     WiringError,
 )
-from .wiring import (
-    Accessor,
-    AioAccessor,
-    Spec,
-    SpecEntry,
-    Wiring,
-)
+from .runtime import Accessor, AioAccessor, Spec, SpecEntry, WiringRuntime
+from .wiring import WiringBase
+
+Wiring = WiringRuntime
 
 __all__ = [
     "Spec",
     "SpecEntry",
     "Wiring",
+    "WiringRuntime",
+    "WiringCompiler",
+    "WiringBase",
     "WiringError",
     "UnknownPlaceholderError",
     "CircularWiringError",
