+++ apywire/wiring.py
@@ -3,22 +3,14 @@
 # SPDX-License-Identifier: ISC
 
 
-"""Core wiring functionality."""
+"""Base wiring functionality."""
 
 from __future__ import annotations
 
-import ast
-import asyncio
-import importlib
 from types import EllipsisType
-from typing import Awaitable, Callable, TypeAlias, cast, final
+from typing import TypeAlias, cast
 
-from apywire.exceptions import (
-    CircularWiringError,
-    UnknownPlaceholderError,
-    WiringError,
-)
-from apywire.thread_safety import CompiledThreadSafeMixin, ThreadLocalState
+from apywire.threads import ThreadLocalState
 
 _ConstantValue: TypeAlias = (
     str | bytes | bool | int | float | complex | EllipsisType | None
@@ -96,20 +88,8 @@ class _ThreadLocalState(ThreadLocalState):
     """
 
 
-# Constant for property AST arguments (used in compile methods)
-_PROPERTY_ARGS = ast.arguments(
-    posonlyargs=[],
-    args=[ast.arg(arg="self")],
-    vararg=None,
-    kwarg=None,
-    defaults=[],
-    kwonlyargs=[],
-    kw_defaults=[],
-)
-
-
-class Wiring(CompiledThreadSafeMixin):
-    """Lazy-loaded container for wired objects."""
+class WiringBase:
+    """Base class for wiring containers."""
 
     _parsed: dict[str, _ParsedEntry]
     _values: dict[str, _RuntimeValue]
@@ -164,7 +144,10 @@ class Wiring(CompiledThreadSafeMixin):
         # Instances will be lazily instantiated and stored in self._values
         if self._thread_safe:
             # Thread-safe mode: use mixin helpers for lock state.
-            self._init_thread_safety(max_lock_attempts, lock_retry_sleep)
+            # Note: Derived classes must implement _init_thread_safety
+            # if they support thread safety (like WiringRuntime).
+            if hasattr(self, "_init_thread_safety"):
+                self._init_thread_safety(max_lock_attempts, lock_retry_sleep)
         else:
             # Non-thread-safe mode: use simple list for resolving stack
             self._resolving_stack: list[str] = []
@@ -207,869 +190,3 @@ class Wiring(CompiledThreadSafeMixin):
         if isinstance(obj, tuple):
             return tuple(self._resolve(v) for v in obj)
         return obj
-
-    def _resolve_runtime(
-        self,
-        o: _ResolvedValue,
-        context: str | None = None,
-    ) -> _RuntimeValue:
-        """Resolve _WiredRef placeholders at runtime.
-
-        This recurses into nested structures and replaces `_WiredRef` with
-        the instantiated `self.<name>` value.
-        """
-        if isinstance(o, _WiredRef):
-            # Detect when a placeholder reference points to something that
-            # wasn't defined in the spec. This is an unknown placeholder
-            # and should raise a friendly error.
-            if o.name not in self._values and o.name not in self._parsed:
-                ctx = f" while instantiating '{context}'" if context else ""
-                raise UnknownPlaceholderError(
-                    f"Unknown placeholder '{o.name}' referenced{ctx}."
-                )
-            # Membership check ensures `o.name` is known; if getattr raises
-            # AttributeError it's from instance creation — let it propagate.
-            # Call the accessor `self.name()` to get the runtime value.
-            return cast(object, getattr(self, o.name)())
-        if isinstance(o, dict):
-            return {
-                k: self._resolve_runtime(v, context=context)
-                for k, v in o.items()
-            }
-        if isinstance(o, list):
-            return [self._resolve_runtime(v, context=context) for v in o]
-        if isinstance(o, tuple):
-            return tuple(self._resolve_runtime(v, context=context) for v in o)
-        return o
-
-    def _astify(self, obj: _ResolvedValue, aio: bool = False) -> ast.expr:
-        """Convert a Python object (possibly a `_WiredRef`) to AST.
-
-        Nested lists, tuples and dicts are supported. `_WiredRef` becomes
-        an accessor call (`self.<name>()`) to mirror runtime behavior.
-
-        If ``aio`` is True, `_WiredRef` becomes an awaited accessor
-        (`await self.<name>()`) to reflect the asynchronous compiled
-        code's behavior.
-        """
-        if isinstance(obj, _WiredRef):
-            # Access the wired value via `self.name()` in compiled code.
-            call = ast.Call(
-                func=ast.Attribute(
-                    value=ast.Name(id="self", ctx=ast.Load()),
-                    attr=obj.name,
-                    ctx=ast.Load(),
-                ),
-                args=[],
-                keywords=[],
-            )
-            if aio:
-                return ast.Await(value=call)
-            return call
-        if (
-            isinstance(obj, (str, bytes, bool, int, float, complex))
-            or obj is None
-            or isinstance(obj, EllipsisType)
-        ):
-            return ast.Constant(obj)
-        if isinstance(obj, dict):
-            keys = [ast.Constant(k) for k in obj.keys()]
-            values = [self._astify(v, aio=aio) for v in obj.values()]
-            return ast.Dict(
-                keys=cast(list[ast.expr | None], keys),
-                values=values,
-            )
-        if isinstance(obj, list):
-            elts = [self._astify(v, aio=aio) for v in obj]
-            return ast.List(elts=elts, ctx=ast.Load())
-        if isinstance(obj, tuple):
-            elts = [self._astify(v, aio=aio) for v in obj]
-            return ast.Tuple(elts=elts, ctx=ast.Load())
-        return ast.Constant(cast(_ConstantValue, obj))
-
-    def _compile_property(
-        self,
-        name: str,
-        module_name: str,
-        class_name: str,
-        data: _ResolvedSpecMapping,
-        *,
-        aio: bool = False,
-        thread_safe: bool = False,
-    ) -> ast.FunctionDef | ast.AsyncFunctionDef:
-        """Build an AST FunctionDef for a cached accessor that returns
-        ``module.class(**data)``.
-
-        When ``aio`` is True this function will produce an
-        ``ast.AsyncFunctionDef`` that awaits referenced accessors and calls
-        the blocking constructor in an executor (``loop.run_in_executor``).
-        When ``aio`` is False it produces a standard synchronous
-        ``ast.FunctionDef``.
-        """
-        module_attr = ast.Attribute(
-            value=ast.Name(id=module_name, ctx=ast.Load()),
-            attr=class_name,
-            ctx=ast.Load(),
-        )
-        kwargs: list[ast.keyword] = []
-        # When compiling async accessors we must precompute awaited
-        # referenced attributes into locals so that the constructor
-        # lambda passed to a thread pool executor is pure synchronous.
-        pre_statements: list[ast.stmt] = []
-
-        # Helper to extract `await self.<name>()` occurrences from AST
-        # expressions and replace them with local variables. This lets us
-        # run a synchronous lambda in `run_in_executor` without `await`
-        # nodes embedded inside it.
-        counter = 0
-
-        def _replace_awaits_with_locals(node: ast.expr) -> ast.expr:
-            nonlocal counter
-
-            if isinstance(node, ast.Await):
-                inner = node.value
-                if (
-                    isinstance(inner, ast.Call)
-                    and isinstance(inner.func, ast.Attribute)
-                    and isinstance(inner.func.value, ast.Name)
-                    and inner.func.value.id == "self"
-                ):
-                    # produce a unique variable name and precompute the await
-                    counter += 1
-                    var_name = f"__val_{counter}"
-                    assign = ast.Assign(
-                        targets=[ast.Name(id=var_name, ctx=ast.Store())],
-                        value=node,
-                    )
-                    pre_statements.append(assign)
-                    return ast.Name(id=var_name, ctx=ast.Load())
-                return node
-
-            # Recurse into common composite nodes
-            if isinstance(node, ast.Call):
-                new_args = [_replace_awaits_with_locals(a) for a in node.args]
-                new_keywords = [
-                    ast.keyword(
-                        arg=k.arg, value=_replace_awaits_with_locals(k.value)
-                    )
-                    for k in node.keywords
-                ]
-                return ast.Call(
-                    func=node.func, args=new_args, keywords=new_keywords
-                )
-            if isinstance(node, ast.Dict):
-                new_keys = [
-                    _replace_awaits_with_locals(k) if k is not None else None
-                    for k in node.keys
-                ]
-                new_values = [
-                    _replace_awaits_with_locals(v) for v in node.values
-                ]
-                return ast.Dict(keys=new_keys, values=new_values)
-            if isinstance(node, ast.List):
-                return ast.List(
-                    elts=[_replace_awaits_with_locals(e) for e in node.elts],
-                    ctx=node.ctx,
-                )
-            if isinstance(node, ast.Tuple):
-                return ast.Tuple(
-                    elts=[_replace_awaits_with_locals(e) for e in node.elts],
-                    ctx=node.ctx,
-                )
-            return node
-
-        args: list[ast.expr] = []
-
-        # Normalize data into args_data (list) and kwargs_data (dict)
-        args_data: list[_ResolvedValue] = []
-        kwargs_data: dict[str, _ResolvedValue] = {}
-
-        if isinstance(data, list):
-            args_data = data
-        else:
-            data_dict = data
-            int_keys = sorted(k for k in data_dict if isinstance(k, int))
-            str_keys = [k for k in data_dict if isinstance(k, str)]
-            for k_int in int_keys:
-                args_data.append(data_dict[k_int])
-            for k_str in str_keys:
-                kwargs_data[k_str] = data_dict[k_str]
-
-        # Process positional args
-        for i, value in enumerate(args_data):
-            raw_val_ast = self._astify(value, aio=aio)
-            if aio:
-                val_ast = _replace_awaits_with_locals(raw_val_ast)
-                var_name = f"__arg_{i}"
-                assign = ast.Assign(
-                    targets=[ast.Name(id=var_name, ctx=ast.Store())],
-                    value=val_ast,
-                )
-                pre_statements.append(assign)
-                arg_val: ast.expr = ast.Name(id=var_name, ctx=ast.Load())
-            else:
-                arg_val = raw_val_ast
-            args.append(arg_val)
-
-        # Process keyword args
-        for key, value in kwargs_data.items():
-            raw_val_ast = self._astify(value, aio=aio)
-            if aio:
-                # Replace any awaited accessors with local precomputed
-                # variables and assign all values to local variables so
-                # that the executor lambda can be synchronous.
-                val_ast = _replace_awaits_with_locals(raw_val_ast)
-                # Use named variables for top-level keyword args to make
-                # the generated code more readable and deterministic.
-                var_name = f"__val_{key}"
-                assign = ast.Assign(
-                    targets=[ast.Name(id=var_name, ctx=ast.Store())],
-                    value=val_ast,
-                )
-                pre_statements.append(assign)
-                kw_val: ast.expr = ast.Name(id=var_name, ctx=ast.Load())
-            else:
-                kw_val = raw_val_ast
-            kwargs.append(ast.keyword(arg=key, value=kw_val))
-
-        # Build the actual constructor call (module.Class(*args, **kwargs))
-        call = ast.Call(func=module_attr, args=args, keywords=kwargs)
-
-        # Cache attribute name like `_name` used to store instantiated
-        # objects on `self` at runtime. The compiled accessor returns
-        # `self._name` when present.
-        cache_attr = f"_{name}"
-
-        # if not hasattr(self, '_name'): ...
-        has_check = ast.UnaryOp(
-            op=ast.Not(),
-            operand=ast.Call(
-                func=ast.Name(id="hasattr", ctx=ast.Load()),
-                args=[
-                    ast.Name(id="self", ctx=ast.Load()),
-                    ast.Constant(value=cache_attr),
-                ],
-                keywords=[],
-            ),
-        )
-
-        return_stmt = ast.Return(
-            value=ast.Attribute(
-                value=ast.Name(id="self", ctx=ast.Load()),
-                attr=cache_attr,
-                ctx=ast.Load(),
-            )
-        )
-        # Build the assignment that sets the cache value; different
-        # behavior is required for async vs sync callers.
-        if not aio:
-            assign_cache = ast.Assign(
-                targets=[
-                    ast.Attribute(
-                        value=ast.Name(id="self", ctx=ast.Load()),
-                        attr=cache_attr,
-                        ctx=ast.Store(),
-                    )
-                ],
-                value=call,
-            )
-        else:
-            # Async path: compute local values and call in executor.
-            loop_assign = ast.Assign(
-                targets=[ast.Name(id="loop", ctx=ast.Store())],
-                value=ast.Call(
-                    func=ast.Attribute(
-                        value=ast.Name(id="asyncio", ctx=ast.Load()),
-                        attr="get_running_loop",
-                        ctx=ast.Load(),
-                    ),
-                    args=[],
-                    keywords=[],
-                ),
-            )
-            # Create a lambda that returns the class call expression
-            # referencing the precomputed locals
-            lambda_expr = ast.Lambda(
-                args=ast.arguments(
-                    posonlyargs=[],
-                    args=[],
-                    vararg=None,
-                    kwarg=None,
-                    defaults=[],
-                    kwonlyargs=[],
-                    kw_defaults=[],
-                ),
-                body=call,
-            )
-            run_call = ast.Await(
-                value=ast.Call(
-                    func=ast.Attribute(
-                        value=ast.Name(id="loop", ctx=ast.Load()),
-                        attr="run_in_executor",
-                        ctx=ast.Load(),
-                    ),
-                    args=[
-                        ast.Constant(value=None),
-                        lambda_expr,
-                    ],
-                    keywords=[],
-                )
-            )
-            assign_cache = ast.Assign(
-                targets=[
-                    ast.Attribute(
-                        value=ast.Name(id="self", ctx=ast.Load()),
-                        attr=cache_attr,
-                        ctx=ast.Store(),
-                    )
-                ],
-                value=run_call,
-            )
-        if_stmt_body = pre_statements.copy()
-        if aio:
-            if_stmt_body.append(loop_assign)
-        if_stmt_body.append(assign_cache)
-        if_stmt = ast.If(test=has_check, body=if_stmt_body, orelse=[])
-        func_def: ast.FunctionDef | ast.AsyncFunctionDef
-        # If the compiled output is thread-safe, inject locking logic
-        if not aio and not thread_safe:
-            func_def = ast.FunctionDef(
-                name=name,
-                args=_PROPERTY_ARGS,
-                body=[if_stmt, return_stmt],
-                decorator_list=[],
-                returns=None,
-                type_comment=None,
-                type_params=[],
-            )
-        elif not aio and thread_safe:
-            # Build sync thread-safe version using code template
-            # Build a lambda maker which is a synchronous callable used by
-            # the helper mixin to invoke the constructor within the
-            # instantiation lock management. We pass this maker to
-            # `self._instantiate_attr(name, maker)`.
-            maker = ast.Lambda(
-                args=ast.arguments(
-                    posonlyargs=[],
-                    args=[],
-                    vararg=None,
-                    kwarg=None,
-                    defaults=[],
-                    kwonlyargs=[],
-                    kw_defaults=[],
-                ),
-                body=call,
-            )
-            call_inst = ast.Call(
-                func=ast.Attribute(
-                    value=ast.Name(id="self", ctx=ast.Load()),
-                    attr="_instantiate_attr",
-                    ctx=ast.Load(),
-                ),
-                args=[ast.Constant(value=name), maker],
-                keywords=[],
-            )
-            assign_cache = ast.Assign(
-                targets=[
-                    ast.Attribute(
-                        value=ast.Name(id="self", ctx=ast.Load()),
-                        attr=cache_attr,
-                        ctx=ast.Store(),
-                    )
-                ],
-                value=call_inst,
-            )
-            if_stmt_body = pre_statements.copy()
-            if_stmt_body.append(assign_cache)
-            func_def = ast.FunctionDef(
-                name=name,
-                args=_PROPERTY_ARGS,
-                body=[
-                    ast.If(test=has_check, body=if_stmt_body, orelse=[]),
-                    return_stmt,
-                ],
-                decorator_list=[],
-                returns=None,
-                type_comment=None,
-                type_params=[],
-            )
-        # This branch handles the synchronous, non-thread-safe case
-        elif aio and not thread_safe:
-            func_def = ast.AsyncFunctionDef(
-                name=name,
-                args=_PROPERTY_ARGS,
-                body=[if_stmt, return_stmt],
-                decorator_list=[],
-                returns=None,
-                type_comment=None,
-                type_params=[],
-            )
-        else:  # aio and thread_safe
-            # Create a maker lambda (synchronous) for the helper
-            # mixin and run it in executor. Precomputed locals are
-            # already present in `pre_statements` to avoid 'await' in
-            # the lambda body.
-            maker = ast.Lambda(
-                args=ast.arguments(
-                    posonlyargs=[],
-                    args=[],
-                    vararg=None,
-                    kwarg=None,
-                    defaults=[],
-                    kwonlyargs=[],
-                    kw_defaults=[],
-                ),
-                body=call,
-            )
-            instantiate_call = ast.Call(
-                func=ast.Attribute(
-                    value=ast.Name(id="self", ctx=ast.Load()),
-                    attr="_instantiate_attr",
-                    ctx=ast.Load(),
-                ),
-                args=[ast.Constant(value=name), maker],
-                keywords=[],
-            )
-            # Build lambda passed to executor that calls the mixin
-            executor_lambda = ast.Lambda(
-                args=ast.arguments(
-                    posonlyargs=[],
-                    args=[],
-                    vararg=None,
-                    kwarg=None,
-                    defaults=[],
-                    kwonlyargs=[],
-                    kw_defaults=[],
-                ),
-                body=instantiate_call,
-            )
-            run_call = ast.Await(
-                value=ast.Call(
-                    func=ast.Attribute(
-                        value=ast.Name(id="loop", ctx=ast.Load()),
-                        attr="run_in_executor",
-                        ctx=ast.Load(),
-                    ),
-                    args=[ast.Constant(value=None), executor_lambda],
-                    keywords=[],
-                )
-            )
-            assign_cache = ast.Assign(
-                targets=[
-                    ast.Attribute(
-                        value=ast.Name(id="self", ctx=ast.Load()),
-                        attr=cache_attr,
-                        ctx=ast.Store(),
-                    )
-                ],
-                value=run_call,
-            )
-            body = pre_statements.copy()
-            # Ensure we compute the event loop variable before calling into
-            # run_in_executor.
-            body.append(loop_assign)
-            body.append(assign_cache)
-            func_def = ast.AsyncFunctionDef(
-                name=name,
-                args=_PROPERTY_ARGS,
-                body=[
-                    ast.If(test=has_check, body=body, orelse=[]),
-                    return_stmt,
-                ],
-                decorator_list=[],
-                returns=None,
-                type_comment=None,
-                type_params=[],
-            )
-        return func_def
-
-    def _compile_constant_property(
-        self,
-        name: str,
-        value: _ConstantValue,
-        *,
-        aio: bool = False,
-    ) -> ast.FunctionDef | ast.AsyncFunctionDef:
-        """Return an AST FunctionDef for an accessor that returns a
-        constant value.
-
-        When ``aio`` is True the compiled accessor will be an
-        ``async def`` that returns the constant directly (no executor
-        required), otherwise a synchronous ``def`` is emitted.
-        """
-        return_stmt = ast.Return(value=ast.Constant(value))
-        func_def: ast.FunctionDef | ast.AsyncFunctionDef
-        if not aio:
-            func_def = ast.FunctionDef(
-                name=name,
-                args=_PROPERTY_ARGS,
-                body=[return_stmt],
-                decorator_list=[],
-                returns=None,
-                type_comment=None,
-                type_params=[],
-            )
-        else:
-            func_def = ast.AsyncFunctionDef(
-                name=name,
-                args=_PROPERTY_ARGS,
-                body=[return_stmt],
-                decorator_list=[],
-                returns=None,
-                type_comment=None,
-                type_params=[],
-            )
-        return func_def
-
-    def __getattr__(self, name: str) -> Callable[[], _RuntimeValue] | Accessor:
-        """Return an accessor callable for `name` that returns the
-        instantiated object for that name when invoked.
-
-        The returned accessor supports synchronous invocation via
-        callable `wired.name()`.
-        Use `await wired.aio.name()` to obtain the instantiated value
-        asynchronously.
-        """
-        # Return cached value (constant or already instantiated)
-        if name in self._values:
-            # Return a callable that returns the current cached value.
-            def _quick() -> _RuntimeValue:
-                return self._values[name]
-
-            return _quick
-
-        if name not in self._parsed:
-            raise AttributeError(f"no attribute '{name}'")
-
-        self._check_circular_dependency(name)
-
-        # Return a callable accessor object that mirrors the previous
-        # synchronous behavior. For asynchronous access use `wired.aio`.
-        return Accessor(self, name)
-
-    @property
-    def aio(self) -> "AioAccessor":
-        """Return a wrapper object providing async accessors.
-
-        Use `await wired.aio.name()` to obtain the instantiated value
-        asynchronously. We use `aio` to avoid the reserved keyword
-        `async` (so `wired.async` would be invalid syntax).
-        """
-        if not hasattr(self, "_aio_accessor"):
-            self._aio_accessor = AioAccessor(self)
-        return self._aio_accessor
-
-    def _access_impl(self, name: str) -> _RuntimeValue:
-        """Common implementation for synchronous accessors.
-
-        This contains the former logic that lived inside the accessor
-        closure and is reused by both the sync call and the async
-        executor wrapper.
-        """
-        # Fast non-thread-safe path
-        if not self._thread_safe:
-            self._get_resolving_stack().append(name)
-            try:
-                try:
-                    return self._instantiate_impl(name)
-                except (UnknownPlaceholderError, CircularWiringError):
-                    raise
-                except WiringError as e:
-                    raise WiringError(f"failed to instantiate '{name}'") from e
-                except Exception as e:
-                    raise WiringError(f"failed to instantiate '{name}'") from e
-            finally:
-                self._get_resolving_stack().pop()
-
-        # Thread-safe mode: use the mixin's `_instantiate_attr` helper.
-        stack = self._get_resolving_stack()
-        stack.append(name)
-        try:
-
-            def _maker() -> _RuntimeValue:
-                return self._instantiate_impl(name)
-
-            return self._instantiate_attr(name, _maker)
-        finally:
-            stack.pop()
-
-    def _check_circular_dependency(self, name: str) -> None:
-        """Check for circular dependencies in the wiring graph.
-
-        Uses either the simple list (non-thread-safe) or thread-local
-        stack to detect cycles.
-        """
-        stack = self._get_resolving_stack()
-        if name in stack:
-            cycle = " -> ".join(stack + [name])
-            raise CircularWiringError(
-                f"Circular wiring dependency detected: {cycle}."
-            )
-
-    def _get_resolving_stack(self) -> list[str]:
-        """Get or initialize the resolving stack for the current mode.
-
-        Returns the per-thread resolving stack when `thread_safe` is True,
-        otherwise returns the single shared list on `self`.
-        """
-        if not self._thread_safe:
-            if not hasattr(self, "_resolving_stack"):
-                self._resolving_stack = []
-            return self._resolving_stack
-
-        # Delegate to mixin for thread-local resolving stack
-        return super()._get_resolving_stack()
-
-    # `_get_held_locks` and `_get_attribute_lock` are provided by the
-    # `CompiledThreadSafeMixin` when `thread_safe` is enabled.
-
-    def _instantiate_impl(self, name: str) -> _RuntimeValue:
-        """Import module, resolve kwargs and instantiate the class."""
-        # If another thread created the instance while we were waiting,
-        # return it to avoid duplicates.
-        if name in self._values:
-            return self._values[name]
-
-        module_name, class_name, ins_data = self._parsed[name]
-        module = importlib.import_module(module_name)
-        cls = cast(  # type: ignore[explicit-any, misc]
-            Callable[..., object], getattr(module, class_name)
-        )
-
-        # Build args and kwargs, resolving wired references to runtime objects
-        args: list[_RuntimeValue] = []
-        kwargs: dict[str, _RuntimeValue] = {}
-
-        if isinstance(ins_data, list):
-            for v in ins_data:
-                args.append(self._resolve_runtime(v, context=name))
-        else:
-            ins_data_dict = ins_data
-            int_keys = sorted(k for k in ins_data_dict if isinstance(k, int))
-            str_keys = [k for k in ins_data_dict if isinstance(k, str)]
-
-            for k_int in int_keys:
-                args.append(
-                    self._resolve_runtime(ins_data_dict[k_int], context=name)
-                )
-
-            for k_str in str_keys:
-                kwargs[k_str] = self._resolve_runtime(
-                    ins_data_dict[k_str], context=name
-                )
-
-        instance = cls(*args, **kwargs)
-        self._values[name] = instance
-        return instance
-
-    # Thread-safety helpers are provided by CompiledThreadSafeMixin and
-    # accessed via `_init_thread_safety()`, `_get_attribute_lock()`,
-    # `_get_held_locks()`, `_release_held_locks()` and
-    # `_instantiate_attr(name, maker)`.
-
-    def compile(self, *, aio: bool = False, thread_safe: bool = False) -> str:
-        """Compiles the Spec into a string containing Python code.
-
-        Args:
-            aio: If True, generate `async def` accessors for wired
-                attributes that await referenced attributes and call
-                blocking constructors in a threadpool via
-                `asyncio.get_running_loop().run_in_executor`. When False
-                (default) generate synchronous `def` accessors.
-
-        Returns:
-            A string containing the Python source for the compiled
-            `Compiled` container.
-        """
-        # Build AST for the module
-        body: list[ast.stmt] = []
-
-        # Add import statements
-        modules = set()
-        for module_name, _, _ in self._parsed.values():
-            modules.add(module_name)
-        if aio:
-            modules.add("asyncio")
-        if thread_safe:
-            # When compiling thread_safe, import thread-safety primitives
-            modules.add("apywire.thread_safety")
-            modules.add("apywire.exceptions")
-        for module in sorted(modules):
-            if module == "apywire.thread_safety":
-                # Import CompiledThreadSafeMixin from thread_safety
-                body.append(
-                    ast.ImportFrom(
-                        module="apywire.thread_safety",
-                        names=[
-                            ast.alias(name="CompiledThreadSafeMixin"),
-                        ],
-                        level=0,
-                    )
-                )
-            elif module == "apywire.exceptions":
-                # Import LockUnavailableError from exceptions
-                body.append(
-                    ast.ImportFrom(
-                        module="apywire.exceptions",
-                        names=[
-                            ast.alias(name="LockUnavailableError"),
-                        ],
-                        level=0,
-                    )
-                )
-            else:
-                body.append(ast.Import(names=[ast.alias(name=module)]))
-
-        # Build class body
-        class_body: list[ast.stmt] = []
-        # When thread safe, add __init__ that calls helper mixin init
-        if thread_safe:
-            # class __init__
-            init_body: list[ast.stmt] = []
-            # compiled class will inherit from the mixin and just call
-            # `_init_thread_safety` from its constructor
-            init_body.append(
-                ast.Expr(
-                    value=ast.Call(
-                        func=ast.Attribute(
-                            value=ast.Name(id="self", ctx=ast.Load()),
-                            attr="_init_thread_safety",
-                            ctx=ast.Load(),
-                        ),
-                        args=[],
-                        keywords=[],
-                    ),
-                )
-            )
-            init_def = ast.FunctionDef(
-                name="__init__",
-                args=_PROPERTY_ARGS,
-                body=init_body,
-                decorator_list=[],
-                returns=None,
-                type_params=[],
-            )
-            class_body.insert(0, init_def)
-        for name, (module_name, class_name, data) in self._parsed.items():
-            class_body.append(
-                self._compile_property(
-                    name,
-                    module_name,
-                    class_name,
-                    data,
-                    aio=aio,
-                    thread_safe=thread_safe,
-                )
-            )
-
-        # Add constant accessors (names present in _values but not in parsed)
-        for name, value in self._values.items():
-            if name in self._parsed:
-                continue
-            class_body.append(
-                self._compile_constant_property(
-                    name,
-                    cast(_ConstantValue, value),
-                    aio=aio,
-                )
-            )
-
-        class_body = [ast.Pass()] if not class_body else class_body
-        # When using thread_safe compiled output we will rely on the
-        # CompiledThreadSafeMixin and _LockUnavailableError imported from
-        # apywire.thread_safety rather than embedding helper code.
-        # Build class definition
-        class_bases: list[ast.expr] = []
-        if thread_safe:
-            class_bases.append(
-                ast.Name(id="CompiledThreadSafeMixin", ctx=ast.Load())
-            )
-        class_def = ast.ClassDef(
-            name="Compiled",
-            bases=class_bases,
-            keywords=[],
-            body=class_body,
-            decorator_list=[],
-            type_params=[],
-        )
-        body.append(class_def)
-
-        # Add compiled = Compiled()
-        assign = ast.Assign(
-            targets=[ast.Name(id="compiled", ctx=ast.Store())],
-            value=ast.Call(
-                func=ast.Name(id="Compiled", ctx=ast.Load()),
-                args=[],
-                keywords=[],
-            ),
-        )
-        body.append(assign)
-
-        # Create module AST
-        module_ast = ast.Module(body=body, type_ignores=[])
-        ast.fix_missing_locations(module_ast)
-
-        # Unparse to string
-        return ast.unparse(module_ast)
-
-
-@final
-class Accessor:
-    """Callable accessor returned by `Wiring.__getattr__`.
-
-    Accessing `wired.name` returns an `Accessor` object which is
-    callable: `wired.name()` returns the instantiated object. If you
-    need async access, use `await wired.aio.name()`.
-    """
-
-    def __init__(self, wiring: Wiring, name: str) -> None:
-        self._wiring = wiring
-        self._name = name
-
-    def __call__(self) -> _RuntimeValue:
-        return self._wiring._access_impl(self._name)
-
-
-@final
-class AioAccessor:
-    """Wrapper used to return async callables via `wired.aio`.
-
-    Example: ``await wired.aio.name()`` will return the requested value
-    asynchronously. If ``name`` is a constant cached in ``_values`` the
-    accessor returns it directly (no executor is used). Otherwise the
-    synchronous instantiation is executed in a threadpool (``run_in_executor``)
-    and the instance is returned.
-    """
-
-    def __init__(self, wiring: Wiring) -> None:
-        self._wiring = wiring
-
-    def __getattr__(
-        self,
-        name: str,
-    ) -> Callable[[], Awaitable[_RuntimeValue]]:
-        # If we already have a cached value (constants and already-created
-        # instances), return an async function that returns it directly.
-        if name in self._wiring._values:
-
-            async def _quick() -> _RuntimeValue:
-                return self._wiring._values[name]
-
-            return _quick
-
-        # Unknown attributes should raise like the sync accessor.
-        if name not in self._wiring._parsed:
-            raise AttributeError(f"no attribute '{name}'")
-
-        async def _acall() -> _RuntimeValue:
-            loop = asyncio.get_running_loop()
-            return await loop.run_in_executor(
-                None, self._wiring._access_impl, name
-            )
-
-        return _acall
-
-
-# `_LockUnavailableError` is imported from `apywire.thread_safety` to
-# keep a single canonical definition shared between runtime and compiled
-# variants. See `apywire/thread_safety.py` for the implementation.
